ARM: mx6-sabrelite: remove obsolete sabrelite_mem_init()
The memory init is done by of_probe_memory() first which is part of the
core_initcall's because the device tree has an valid memory node.
Otherwise the memory-controller driver (esdctl) would add it by reading
the information from the controller which are set by the DCD data.
Therefore we can drop it from the local board file.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>