ARCH: i.MX6: add support for OFDEVICE
The imx23_defconfig, imx_v8_defconfig and imx_defconfig already
selecting this config except for the imx_v7_defconfig. The
imx_v7_defconfig selects MACH_ZII_VF610_DEV which selects ARCH_VF610
which select OFDEVICE. So it will be never selected if the
MACH_ZII_VF610_DEV board support is dropped.

Selecting the option here seems to be the better place instead of adding
it to the imx_v7_defconfig because there are possible no i.MX6 boards not
using the device tree.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>