ARM: protonic: disable on-die termination to fix PHY bootstrapping
If on-die termination is enabled, the RXC pin of iMX6 will be pulled
high. Since we already have an 10K pull-down on board, the RXC level on
PHY reset will be ~800mV, which is mostly interpreted as 1. On some
reboots we get 0 instead and kernel can't detect the PHY properly.

Since the default 0x020e07ac value is 0, it is sufficient to remove this
entry from the affected imxcfg files.

Since we get stable 0 on pin PHYADDR[2], the PHY address is changed from
4 to 0.

Fixes: 00adc1e33ef8 ("ARM: add imx6 based Protonic boads")
Signed-off-by: Oleksij Rempel <o.rempel@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>