sandbox: hostfile: maintain created temp files over reset
Currently, hostfils without a barebox,filename are assigned a new
temporary file on every start.

For use of a hostfile syscon for reboot mode or reset source, we will
need to persist the changes over a regular reset. This can easily be
done by passing through all automatically created temporary file with
an explicit --image option on subsequent runs. Do so.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>