commands: bootchooser: migrate menu entry to boot menu
While the bootchooser command is not required for boot with bootchooser,
it's still boot related and is more likely to be looked for in the boot
menu than the misc menu. The boot menu has other commands like uimage
which do not boot either, but are boot-related.

While at it, add some Kconfig help text.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>