mtd: nand-imx: repair reading the oob area
After sending the READOOB command the core (in nand_read_oob_std()) uses
the read_buf callback to fetch the result. So in the imx driver it is
necessary to emit the NFC_OUTPUT command to make the controller actually
read the requested data into the internal buffer and then copy the data
to the location where read_buf expects it.

This fixes reading from /dev/nand0.oob and also (somewhat) reading bad
block markers in the absence of a bad block table.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Tested-by: Sascha Hauer <s.hauer@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>