doc: add note about characters like hyphens in variable names
The mismatch between the allowed characters in shell variables and
device parameters is a common pitfall, especially because device
tree node names often have hyphens in them, which is disallowed by
hush.

While such variables can be read by the ${variable-example}, setting
them is only possible with the setenv command. Reflect this in the
documentation.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>