i.MX: HABv4: Improve HAB event printing
Instead of using a fixed sized buffer for the report_event function,
let's call it two times, once with a NULL pointer to get the size of the
event and a second time with a buffer of that size.
Also, instead of separating the events into warning and error type,
iterate over all events in one single loop. This helps to get the events
in the order they occured which probably helps the reader to make more
sense of them.
This also gets rid of the "ERROR: Recompile with larger event data buffer"
message which recently appeared everytime a warning or error was
reported.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>