Revert "ARM: dts: bcm2835-rpi: re-enable booting from SD card"
Since d6d3d0aecece ("mci: add driver for BCM283x sdhost core"), we now
have driver support for the sdhost controller in barebox as well.
We can thus drop the device tree override and use the same MCI the
kernel uses. This was tested on a Raspberry Pi zero W, which has
the same bcm2835 SoC as the Raspberry Pi 1.

This reverts commit abc0447440d68af720127fcca3c17f04ee144fcc.

Cc: Roland Hieber <rhi@pengutronix.de>
Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>