ARM: rpi: add new init function for Raspberry Pi Zero
The Raspberry Pi Zero and Zero W have no power LED and lack the USB
hub and ethernet port. The rpi_b_plus_init is thus the wrong init function
to use. Add a new init function that just configures the activity LED
and configure the new exposed USB port for OTG.

While we have a dwc2 gadget driver, it doesn't yet work for the Rpi:

	barebox@RaspberryPi Zero W:/ usbgadget -a
	udc0: registering UDC driver [g_multi]
	multi_bind: creating ACM function
	g_multi usbgadget: Multifunction Composite Gadget
	g_multi usbgadget: g_multi ready
	dwc2 20980000.usb@7e980000.of: bound driver g_multi
	WARNING: at drivers/usb/dwc2/gadget.c:1896/dwc2_gadget_pullup()!

But let's add dr_mode = otg anyway, so it can be debugged out of the box
with only config changes.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>