doc: bcm283x: document use of mini-uart on Raspberry Pi Zero W / CM3
As described in the UART configuration[1] article in the Raspberry Pi
Foundation documentation, Raspberry Pi 3 & 4 as well as Zero W use the
mini-uart as primary (easily user-accessible) UART. At least on the
Raspberry Zero W and CM3, we need to pass uart_2ndstage=1, so the
BootROM leaves the 8250 IP in a suitable state for use by barebox.
Document this.

[1]: https://www.raspberrypi.org/documentation/configuration/uart.md

Cc: Roland Hieber <rhi@pengutronix.de>
Cc: Rouven Czerwinski <rcz@pengutronix.de>
Cc: Robert Carnecky <robert@neopsis.com>
Cc: Andrew John <andrew@mobius-embedded-systems.co.uk>
Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>