include: <linux/math64.h>: sync with upstream
The header implements definitions for the 64-bit division helpers
on 64-bit builds only. For 32-bit builds, it can only provide prototypes
and the actual implementation will need to come from elsewhere.

We didn't have any out-of-line definitions in barebox with the result
that functions like div_s64_rem() were so far only usable in
64-bit barebox builds. On 32-bit builds, they would result in a linker
error.

Import the Linux v5.11-rc1 generic out-of-line 64-bit math on 32-bit
implementation to fix this. While at it, synchronize the header to
reduce diff to upstream.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>