ARM: at91: at91sam9x5: fix co-existance of erratum-aware and generic reset
We have a generic at91sam9 reset driver, but it's unaware of the
erratum on the at91sam9x5, which can prevent reboot from NAND due to
interference from SDRAM. The workaround is packing the powering down
of the DDR and the system reset into a single cache line and executing
that. This would be a bit tedious to add into the device tree probed
driver, thus:

 - Don't activate the work around if we are device-tree enabled, but
   have a newer SoC
 - Give the workaround a slightly higher priority, so it's taken instead
   of the generic DT driver

This fixes an issue of failing reset with the at91_multi_defconfig,
because both reset drivers have the same priority of 100.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>