ARM: boards: kontron-samx6i: fix memory detection helper
According the vendor u-boot version [1] the version pin is not involved
in the memory size detection algorithm. So drop the usage here.

Also fix the id0 gpio request. Instead of requesting GPIO6_6 we need to
request GPIO6_7.

Last we need to fix the size for the solo/duallite modules. Those
modules are populated with at least 256M of RAM according [1].

I fixed also the alignment while on it.

[1] https://github.com/kontron/u-boot-smarc-samx7/blob/master/board/kontron/amx6/amx6.c

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>