ARM: dts: imx6qdl-samx6i: use upstream device tree files
Make use of upstream available device tree files to avoid duplications.
A few interfaces had a "default on" behaviour which is dropped now.
Enabling those interfaces within the SoM dtsi seems to be wrong anyway.
Since boards have to turn it off again if those interfaces are unused.

I've adapted all upstream boards which are using the the
imx6qdl-smarc-samx6i.dtsi so the behaviour for those boards is
unchanged.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>