Documentation: state: recommend a partition type GUID for state GPT partitions
The partition type GUID was generated randomly. Having a defined GUID
should reduce the risk of collisions in the future.

A dedicated GUID could also allow us to implement a fallback in case no
state backend property is defined. This could then look for a GPT
partition with this partition type GUID and use that as the backend.
dt-utils could do the same in userspace. This would be similar to [1].

[1] https://www.freedesktop.org/wiki/Specifications/DiscoverablePartitionsSpec/

Signed-off-by: Bastian Krause <bst@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>