efi: declare prototype for close_protocol BootService
Sometimes, drivers need to probe whether the EFI device in front of them
is the actual hardware they expect. If it's not, the driver should call
close_protocol to leave the decision on whether to match this device to
another driver. Use the correct type for close_protocol to facilitate
this.

No functional change, as no in-tree driver makes use of this yet.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>