watchdog: Print seconds to expire
This adds a parameter to the watchdog devices that shows what we think
when the watchdog expires. The watchdog should reset the system once the
counter hits zero. When the system resets earlier or the counter shows
negative values then there might be problems with the watchdog. Useful
for debugging watchdog related problems.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>