watchdog: stpmic1-wdt: don't modify global turn-off behaviour in probe
According to Table 12 in the STPMIC1-datasheet (Rev 7 - December 2020)
there are the following turn-OFF conditions:

- Software switch-OFF*
- PONKEYn long key press*
- Thermal shutdown
- Overcurrent protection
- Watchdog*
- VIN_OK_Fall*

The conditions marked with an * can be re-configured from turn-Off conditions
to restart requests using the RREQ_EN-bit in the CR register
(Called RESTART_REQUEST_ENABLED in Linux/Barebox).

The watchdog code enabled this bit in it's probe function to trigger a restart
on a watchdog event. This may however result in unexpected behaviour,
like a long-press of the PONKEY resulting in a restart instead of a turn-off
when the watchdog is enabled vs. when it is not.

This change does not affect the restart and poweroff functionality provided by
the wdt driver, as the restart-/poweroff-handlers explicitly set/clear the
RESTART_REQUEST_ENABLED flag when triggering a software turn-OFF.

To prevent unexpected behaviour of the PONKEYn and other turn-OFF sources the
RESTART_REQUEST_ENABLED configuration in the probe function can and should thus be removed.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>