fs: increase reference count for backing store when loop mounting
The VFS layer already increase the mount reference count for the mount
point. This means e.g. following sequence is well-behaving:

 mkdir -p /mnt/disk0.0/media
 mount -o loop backing.squashfs /mnt/disk0.0/media
 umount /mnt/disk0.0

barebox will do the right thing and report

  umount: Device or resource busy

However the reference count of the file where backing.squashfs comes
from is not incremented on mount with the effect that following sequence
crashes:

 mkdir -p /media
 mount -o loop /mnt/disk0.0/backing.squashfs /media
 umount /mnt/disk0.0 # should've returned EBUSY
 umount /media

Fix this by touching the backing store's mount reference count when
loop mounting and unmounting.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>