video: add simplefb driver
barebox has support to fix up a framebuffer it has set up as simplefb
into the device tree of a kernel it boots. Add the counterpart to this,
so barebox itself can reuse an already set up frame buffer.

This is done to support the framebuffer device on the tinyemu RISC-V
machine.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>